run/animation/frontend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
requests==2.31.0
google-cloud-storage==2.9.0
gunicorn==21.2.0
